Attempt to get GitHub up to date again! Touch not tested through at present

+++ Library/Dialog/MonoDroid.Dialog/EntryElement.cs
@@ -90,7 +90,7 @@ namespace MonoDroid.Dialog
                 else if (Send != null)
                 {
                     _entry.ImeOptions = ImeAction.Go;
-                    _entry.SetImeActionLabel("Go", (int)ImeAction.Go);
+                    _entry.SetImeActionLabel("Go", ImeAction.Go);
                     _entry.EditorAction = (o, actionId, e) =>
                     {
                         if (actionId == ImeAction.Go)
